Add service worker implementation
* Objective

This commit objective is to add the service worker implemenation that
should be deployed to capture dApp events.

* Why

A service worker is the approach used by encinitas to capture the dApp
events, the service worker get installed by the initialization script
which starts tracking all the events.

* How

This commit adds the service worker implementation